rooms
Added another room, w_slime_forest_rm2. Also updated the sprites for the spider. Both the spider and the web that it shoots.
Also updated the bird sprite so it looks really nice and good. Though it still needs the added health, because it hits and dies against the wall way too fast.